Refactor: add UserTokenActivationDisplay to mobile display
